// functions/src/services/dto/search.dto.ts
export interface SearchDto {
    profileId: string;
    petId: string;
    perType: string;
    imageProfile: string;
    imagePet: string;
    profileName: string;
    petName: string;
    age: number;
    petGender: string;
    city: string;
    petPhotos: string[];
    distance: string;
    certificateCount: number;
}